+++ src/types/theme/BaseLineHeight.ts
@@ -6,5 +6,6 @@ export type BaseLineHeight = {
     tight: number
     normal: number
     relaxed: number
+    loose: number
 }