+++ src/types/theme/BaseZIndex.ts
@@ -5,8 +5,11 @@ export type BaseZIndex = {
     behind: number
     base: number
     dropdown: number
+    sticky: number
     overlay: number
     modal: number
+    popover: number
     tooltip: number
+    toast: number
 }